requirements.txt
fastapi==0.111.1
fastapi-utils==0.2.1
uvicorn==0.22.0
redis==4.6.0
PyMySQL==1.1.0
loguru==0.7.0
python-dotenv==1.0.0
SQLAlchemy==1.4.49
grpcio==1.56.2
grpcio-tools==1.56.2
openpyxl==3.1.2
python-multipart==0.0.7
munch==4.0.0
requests==2.31.0
urllib3==2.0.4
aiomysql==0.2.0
pandas==2.1.4
wrapt==1.16.0
celery==5.3.6
jinja2==3.1.2
PyJWT==2.8.0
werkzeug==2.3.7
oss2==2.18.4
xlsxwriter==3.2.0
user-agents==2.2.0
eventlet==0.36.1
python-consul==1.1.0
netifaces2==0.0.22
ulid-py==1.1.0
